Treat global headers as regular files
Generally they can be extracted as regular files -- they just don't
mean a whole lot. This handles cases in which a "pax_global_header"
file is included in the archive.

I'm not entirely sure of the ramifications of this change, but I
couldn't figure out a graceful way to just _ignore_ the file, so
this will have to do for now.